Movie Website/requirements.txt
Flask==3.0.3
imdbpy==2023.05.01
